+++ src/main/java/com/douane/managed/bean/JasperData/OrdreEntreeData.java
@@ -18,15 +18,15 @@ public class OrdreEntreeData {
 		OrdreEntree oe;
 		for (Object[] des : this.data) {
 			Nomenclature nom = (Nomenclature) des[0];
-			Float f = (Float) des[1];
+			Double f = (Double) des[1];
 			List <Object[]> des2 = (List<Object[]>) des[2];
 			for (Object[] desu : des2) {
 				Designation desu0 = (Designation) desu[0];
 				String designat = desu0.getTypematerieladd().getDesignation() + " - " + desu0.getMarque().getDesignation() + " - " + desu0.getRenseignement();
 				String esp = desu0.getEspeceUnite();
 				Long lo = (Long) desu[1];
-				Float pu = desu0.getPu();
-				Float som = (Float) des[1];
+				Double pu = desu0.getPu();
+				Double som = (Double) des[1];
 				String orig = desu0.getRefFacture();
 				oe = new OrdreEntree(nom.getNomenclature(), designat, esp, lo, pu,orig, som);
 				newOne.add(oe);
@@ -43,10 +43,10 @@ public class OrdreEntreeData {
 		private String designations;
 		private String especeUnite;
 		private Long nbr;
-		private Float pu;
-		private Float sommaire;
+		private Double pu;
+		private Double sommaire;
 		private String origine;
-		public OrdreEntree(String n, String d, String e, Long nb,Float p, String o, Float som) {
+		public OrdreEntree(String n, String d, String e, Long nb,Double p, String o, Double som) {
 			this.nomenclature = n;
 			this.designations = d;
 			this.especeUnite = e;
@@ -79,10 +79,10 @@ public class OrdreEntreeData {
 		public void setNbr(Long nbr) {
 			this.nbr = nbr;
 		}
-		public Float getPu() {
+		public Double getPu() {
 			return pu;
 		}
-		public void setPu(Float pu) {
+		public void setPu(Double pu) {
 			this.pu = pu;
 		}
 		public String getOrigine() {
@@ -91,10 +91,10 @@ public class OrdreEntreeData {
 		public void setOrigine(String origine) {
 			this.origine = origine;
 		}
-		public Float getSommaire() {
+		public Double getSommaire() {
 			return sommaire;
 		}
-		public void setSommaire(Float sommaire) {
+		public void setSommaire(Double sommaire) {
 			this.sommaire = sommaire;
 		}
 	}
